pyproject.toml
[tool.poetry]
name = "quantitizer"
version = "1.1.0"
description = "This package allows to quantitize matrix to compress different embedding models."
authors = ["CapBlood <stalker.anonim@mail.ru>"]
homepage = "https://github.com/CapBlood/quantitizer"
#readme = "README.md"
license = "GPL-3.0-only"
#repository = "https://github.com/CapBlood/quantitizer"

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
numpy = "^1.23.1"
scipy = "^1.8.1"
gensim = "^4.2.0"
"mega.py" = "^1.0.8"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
numpydoc = "^1.4.0"
Sphinx = "^5.0.2"
sphinx-rtd-theme = "^1.0.0"
sphinxcontrib-napoleon = "^0.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
